requirements.txt
fastapi
uvicorn[standard]
scikit-learn
spacy
pydantic
python-multipart
pdfplumber
python-docx
numpy
pandas
jinja2
python-dotenv
